Actualización CU-Modificar plan de trabajo

+++ NetBeansProjects/SistemaCuerpoAcademico/src/GUI/WorkPlanActionModifyController.java
@@ -10,8 +10,11 @@ import businessLogic.GoalDao;
 import domain.Action;
 import domain.Goal;
 import domain.Member;
+import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.ResourceBundle;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -21,14 +24,17 @@ import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
 import javafx.scene.control.ComboBox;
+import javafx.scene.control.DatePicker;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.input.MouseEvent;
 import javafx.stage.Stage;
+import javafx.util.converter.LocalDateStringConverter;
 import log.BusinessException;
 import log.Log;
 
@@ -60,8 +66,6 @@ public class WorkPlanActionModifyController implements Initializable {
     @FXML
     private TextField tfResponsable;
     @FXML
-    private TextField tfDate;
-    @FXML
     private TextField tfResource;
     @FXML
     private Button btAdd;
@@ -78,6 +82,8 @@ public class WorkPlanActionModifyController implements Initializable {
     private Button btUpdate;
     @FXML
     private Button btFinish;
+    @FXML
+    private DatePicker dpDateEnd;
     
     /**
      * Initializes the controller class.
@@ -85,12 +91,14 @@ public class WorkPlanActionModifyController implements Initializable {
     @Override
     public void initialize(URL url, ResourceBundle rb) {
         tcAction.setCellValueFactory(new PropertyValueFactory<Action,String>("description"));
-        tcDate.setCellValueFactory(new PropertyValueFactory<Action,String>("dateFinish"));
+        tcDate.setCellValueFactory(new PropertyValueFactory<Action,String>("dateEnd"));
         tcResponsable.setCellValueFactory(new PropertyValueFactory<Action,String>("memberInCharge"));
         tcResource.setCellValueFactory(new PropertyValueFactory<Action,String>("resource"));
         cbGoals.setOnAction(e -> {
             clicOnGoal();    
         });
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
+        dpDateEnd.setConverter(new LocalDateStringConverter(formatter, null));
     }    
     
     public void setMember(Member member) {
@@ -115,10 +123,11 @@ public class WorkPlanActionModifyController implements Initializable {
     private void initializeActions() {
         ActionDAO actionDAO = new ActionDAO();
         try{
-            actions = FXCollections.observableArrayList(actionDAO.getActionsByGoalId(idGoalSelected));
-            oldActions = FXCollections.observableArrayList();
-            oldActions = actions;
-            tvActions.setItems(actions);
+            actions = FXCollections.observableArrayList(actionDAO.getActionsByGoalId(idGoalSelected));  
+            oldActions = FXCollections.observableArrayList(actionDAO.getActionsByGoalId(idGoalSelected));  
+            if(actions != null){
+                tvActions.setItems(actions);   
+            }
         }catch(BusinessException ex){
             Log.logException(ex);
         }
@@ -140,30 +149,40 @@ public class WorkPlanActionModifyController implements Initializable {
     private void actionSave(ActionEvent event) {
         ActionDAO actionDAO = new ActionDAO();
         try{
-            for (int i = 0; i < oldActions.size(); i++){
-                actionDAO.deletedActionById(oldActions.get(i).getId());
+            if(oldActions != null){
+                for (int i = 0; i < oldActions.size(); i++){
+                    actionDAO.deletedActionById(oldActions.get(i).getId());
+                }
             }
             
-            for (int i = 0; i < actions.size(); i++){
-                //MANDAR A LLAMAR A MÉTODO PARA GUARDAR ACCIONES DE LA META
-            }
-            
+            if(actions != null){
+                for (int i = 0; i < actions.size(); i++){
+                    actionDAO.saveSuccesful(actions.get(i), idGoalSelected);
+                }     
+            } 
+            AlertMessage alertMessage = new AlertMessage();
+            alertMessage.showAlertSuccesfulSave("Las acciones  ");
+            actions.clear();
+            oldActions.clear();
         }catch (BusinessException ex) {
             Log.logException(ex);
         }
+        
     }
 
     @FXML
     private void actionAddAction(ActionEvent event) {
         String description = tfAction.getText();
         String responsable = tfResponsable.getText();
-        String date = tfDate.getText();
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
+        String date= dpDateEnd.getValue().format(formatter);
         String resource = tfResource.getText();
         Action action = new Action(description, responsable, date, resource);
-        //MANDAR A VALIDAR
-        actions.add(action);
-        tvActions.refresh();
-        cleanFields();
+        if(validateAction(action)){
+            actions.add(action);
+            tvActions.refresh();
+            cleanFields();
+        }
     }
 
     @FXML
@@ -173,6 +192,9 @@ public class WorkPlanActionModifyController implements Initializable {
             actions.remove(action);
             tvActions.refresh();
             cleanFields();
+        }else{
+            AlertMessage alertMessage = new AlertMessage();
+            alertMessage.showAlertValidateFailed("Selección de acción faltante");
         }
     }
     
@@ -183,16 +205,18 @@ public class WorkPlanActionModifyController implements Initializable {
             int indexAction = actions.indexOf(action);
             String description = tfAction.getText();
             String responsable = tfResponsable.getText();
-            String date = tfDate.getText();
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
+            String date= dpDateEnd.getValue().format(formatter);
             String resource = tfResource.getText();
-             //Mandar a validar que no sea repetido y settear
-            action.setDescription(description);
-            action.setMemberInCharge(responsable);
-            action.setDateFinish(date);
-            action.setResource(resource);
-            actions.set(indexAction, action);
-            tvActions.refresh();
-            cleanFields();
+            Action newAction = new Action(description, responsable, date, resource);
+            if(validateAction(newAction)){
+                actions.set(indexAction, newAction);
+                tvActions.refresh();
+                cleanFields();
+            }
+        }else{
+            AlertMessage alertMessage = new AlertMessage();
+            alertMessage.showAlertValidateFailed("Selección de acción faltante");
         }
     }
 
@@ -219,7 +243,9 @@ public class WorkPlanActionModifyController implements Initializable {
         if (action != null){
             tfAction.setText(action.getDescription());
             tfResponsable.setText(action.getMemberInCharge());
-            tfDate.setText(action.getDateFinish());
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
+            LocalDate date = LocalDate.parse(action.getDateEnd(), formatter);
+            dpDateEnd.setValue(date);
             tfResource.setText(action.getResource());
         }
     }
@@ -227,11 +253,70 @@ public class WorkPlanActionModifyController implements Initializable {
     private void cleanFields(){
         tfAction.setText("");
         tfResponsable.setText("");
-        tfDate.setText("");
+        dpDateEnd.setValue(null);
         tfResource.setText("");
     }
 
     @FXML
     private void actionFinish(ActionEvent event) {
+        Stage stage = (Stage) btFinish.getScene().getWindow();
+        stage.close();
+        openMenu();
+    }
+        
+    private void openMenu(){
+        Stage primaryStage= new Stage();
+            try{
+                URL url = new File("src/GUI/Menu.fxml").toURI().toURL();
+                FXMLLoader loader = new FXMLLoader(url);
+                loader.setLocation(url);
+                loader.load();
+                MenuController menu = loader.getController();
+                menu.initializeMenu(member);
+                Parent root = loader.getRoot();
+                Scene scene = new Scene(root);
+                primaryStage.setScene(scene);
+                primaryStage.show();
+            }catch (IOException ex) {
+                Log.logException(ex);
+            }
+    }
+    
+    private boolean validateAction(Action action){ 
+        boolean value=true;
+        Validation validation = new Validation();
+        AlertMessage alertMessage = new AlertMessage();
+            if(action.getDescription().isEmpty() 
+              ||  action.getMemberInCharge().isEmpty() || action.getResource().isEmpty()
+            ){  
+                value=false;
+                alertMessage.showAlertValidateFailed("Campos vacios");
+            }
+            
+            if(validation.findInvalidField(action.getDescription())
+              || validation.findInvalidField(action.getResource())
+              || validation.findInvalidField(action.getMemberInCharge())
+            ){   
+                value=false;
+               alertMessage.showAlertValidateFailed("Campos invalidos");
+            } 
+            
+            if(repeatedAction(action)){ 
+               value=false;
+               alertMessage.showAlertValidateFailed("Acción repetida");
+            }
+        return value;
+    }
+    
+    public boolean repeatedAction(Action action){
+        Boolean value = false;
+        int i = 0;
+        while(!value && i<actions.size()){
+            if(actions.get(i).equals(action)){
+                value = true;
+            }
+            i++;
+        }
+       return value;
     }
 }
